requirements.txt
django-countries
django-money
weasyprint
